cryptpilot-convert: add a 'local_install_package' option
In offline environments or when using an ISO that does not include the cryptpilot,
 attestation-agent, or confidential-data-hub packages,
the installation fails because these packages are not able to be downloaded from the YUM repository.

To address this, add an option to install these packages from local RPM files.

When using this option, users must manually resolve all dependencies.